authorizer/requirements.txt
django
djangorestframework
djangorestframework-jsonapi
djangorestframework-simplejwt
django-filter
pyyaml
flake8
boto3
